requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
boto3==1.42.21
botocore==1.42.21
click==8.3.1
ecdsa==0.19.1
fastapi==0.128.0
h11==0.16.0
idna==3.11
jmespath==1.0.1
passlib==1.7.4
psycopg2-binary==2.9.11
pyasn1==0.6.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-jose==3.5.0
python-multipart==0.0.21
rsa==4.9.1
s3transfer==0.16.0
six==1.17.0
SQLAlchemy==2.0.45
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.40.0

background-tasks
pydantic-settings